+++ product-service/src/main/java/com/example/product/dto/admin/AdminProductListDto.java
@@ -3,13 +3,15 @@ package com.example.product.dto.admin;
 import lombok.Builder;
 import lombok.Getter;
 
+import java.math.BigDecimal;
+
 @Getter
 @Builder
 public class AdminProductListDto {
 
     private String imageUrl;   // /images/menu/{menuCode}.jpg
-    private Integer menuCode;
+    private Long menuCode;
     private String menuName;
     private String category;
-    private Integer basePrice;
+    private BigDecimal basePrice;
 }
