+++ product-service/src/main/java/com/example/product/dto/user/MenuDetailDto.java
@@ -3,18 +3,19 @@ package com.example.product.dto.user;
 import lombok.Builder;
 import lombok.Getter;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 @Getter
 @Builder
 public class MenuDetailDto {
 
-    private Integer menuCode;
+    private Long menuCode;
     private String imageUrl;
     private String menuName;
     private String description;
     private String category;
-    private Integer basePrice;
+    private BigDecimal basePrice;
     private String baseVolume;
 
     private List<AllergyDto> allergies;   // allergy_ids 파싱 결과
