+++ product-service/src/main/java/com/example/product/dto/user/MenuListDto.java
@@ -6,7 +6,7 @@ import lombok.Getter;
 @Getter @Builder
 public class MenuListDto {
 
-    private Integer menuCode;
+    private Long menuCode;
     private String imageUrl; // /images/menu/{menuCode}.jpg
     private String menuName;
     private String description;
