+++ product-service/src/main/java/com/example/product/dto/user/OptionValueDto.java
@@ -3,11 +3,13 @@ package com.example.product.dto.user;
 import lombok.Builder;
 import lombok.Getter;
 
+import java.math.BigDecimal;
+
 @Getter
 @Builder
 public class OptionValueDto {
 
     private Integer optionId;
     private String optionName;
-    private Integer price;
+    private BigDecimal price;
 }
