+++ product-service/src/main/java/com/example/product/model/Menu.java
@@ -3,6 +3,8 @@ package com.example.product.model;
 import jakarta.persistence.*;
 import lombok.*;
 
+import java.math.BigDecimal;
+
 @Entity
 @Table(name = "menu")
 @Getter @Setter
@@ -14,7 +16,7 @@ public class Menu {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "menu_code")
-    private Integer menuCode;
+    private Long menuCode;
 
     @Column(name = "menu_name", nullable = false, unique = true)
     private String menuName;
@@ -26,7 +28,7 @@ public class Menu {
     private String category;
 
     @Column(name = "base_price", nullable = false)
-    private Integer basePrice;
+    private BigDecimal basePrice;
 
     @Column(name = "base_volume")
     private String baseVolume;
