+++ product-service/src/main/java/com/example/product/repository/MenuRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MenuRepository extends JpaRepository<Menu, Integer> {
+public interface MenuRepository extends JpaRepository<Menu, Long> {
 
     // 1. 전체 메뉴 조회
     Page<Menu> findAll(Pageable pageable);
